requirements.txt
pyautogen
python-dotenv
litellm
litellm[proxy]